src/requirements.txt
asgiref==3.6.0
colorama==0.4.6
coverage==7.2.3
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
flake8==5.0.4
flake8-html==0.4.3
importlib-metadata==4.2.0
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.6
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.15.0
PyJWT==2.6.0
pytest==7.3.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
zipp==3.15.0
